Working solution without stretch goal.
