+++ week2/backend/package.json
@@ -4,7 +4,7 @@
   "description": "Backend for Week 2 assignment",
   "main": "service.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
     "start": "nodemon service.js"
   },
   "repository": {
