--- week2/backend/routes/index.js
@@ -1,6 +1,0 @@
-const routes = require('express').Router();
-const controller = require('../controllers/');
-
-routes.get('/professional', controller.getProfessional);
-
-module.exports = routes;
