+++ week2/backend/routes/professional.js
@@ -1,0 +1,6 @@
+const routes = require('express').Router();
+const controller = require('../controllers/professional');
+
+routes.get('/', controller.getProfessional);
+
+module.exports = routes;
